Split EntityListSystem.kts into separate scripts
