+++ engine/src/main/kotlin/rs/dusk/world/entity/npc/RandomWalking.kts
@@ -15,7 +15,7 @@ Tick then {
         val walkMask = npc.def.walkMask.toInt()
         if (walkMask and 0x1 != 0) {
             if (walkMask and 0x2 != 0 && 100.0 * Math.random() < 10.0) {
-                val randomX = (10.0 * Math.random() - 5.0).roundToInt()
+                val randomX = (Math.random() * 10.0 - 5.0).roundToInt()
                 val randomY = (Math.random() * 10.0 - 5.0).roundToInt()
                 if (randomX != 0 || randomY != 0) {
                     pf.find(npc, npc.tile.add(randomX, randomY))
